Linear2CMap code split and refactoring.
