+++ main.py
@@ -2,7 +2,7 @@ from view.window import Game
 
 import sys
 import json
-import model.Linear2CMap
+import model.mesh_struct.mesh as mesh
 
 # Press the green button in the gutter to run the script.
 if __name__ == '__main__':
@@ -12,6 +12,6 @@ if __name__ == '__main__':
     else:
         f = open(sys.argv[1])
         json_mesh = json.load(f)
-        cmap = model.Linear2CMap.Mesh(json_mesh['nodes'], json_mesh['faces'])
+        cmap = mesh.Mesh(json_mesh['nodes'], json_mesh['faces'])
         g = Game(cmap)
         g.run()
